arm implementation
This adds an NEON version and a version for ARMv6 DSP instructions of Adler32.

The NEON Version is coded in intrinsic, but there are some details
left with the endianness. If i understand it right it should work with
any ARM compiler which understands NEON.
The ARMv6 DSP version uses inline ASM, so i restricted it to GCC (or
compatible compiler, like clang, can also be used. Clang should find
the way in by itself, it disguises as GCC 4.2)

Unfortunately i don't have any ARM, so the Code is untested, and there
are some TODO left.
Also some benchmarking would be needed.