fix PPC Altivec version
The PIM says that after a vec_lde the rest of the vector is undefined.
On a G4 you get the old content of the register.
On a G5 vec_lde behaives like a simple load, the complete vector is
filled with the data surrounding the element you want to load.

So add code to load the one element we really want.