Fix ia64 speculative loads
Speculative loads do not generate any faults, as intendet. Instead
you should check the valid bit of your speculative load.

But speculative loads are great to shedule your loop. If you have
another loop iteration, your load of the next iteration data was
"good", if not, who cares. No need to check any bits.

The problem is: Speculative loads do not generate _any_ faults
(Wait, we had that...). So they also do not generate any valid
page faults to signal the OS action is required because a page
is swapped out or a mmaped file page needs to be fetched.

So fetch the buffer end of this outer loop iteration at each
outer loop start.

And maybe we even have to be more agressive about that...